+++ backend/team6/src/main/java/programmers/team6/domain/admin/repository/AdminVacationRequestSearchCustom.java
@@ -9,9 +9,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 import programmers.team6.domain.admin.dto.AdminVacationSearchCondition;
 import programmers.team6.domain.admin.dto.VacationRequestSearchResponse;
-import programmers.team6.domain.admin.utils.CriteriaCustomPredicateBuilder;
-import programmers.team6.domain.admin.utils.CriteriaCustomQueryBuilder;
-import programmers.team6.domain.admin.utils.QueryUtils;
+import programmers.team6.global.querybuilder.CriteriaCustomPredicateBuilder;
+import programmers.team6.global.querybuilder.CriteriaCustomQueryBuilder;
+import programmers.team6.global.querybuilder.QueryUtils;
 import programmers.team6.domain.member.entity.Code_;
 import programmers.team6.domain.member.entity.Dept_;
 import programmers.team6.domain.member.entity.Member_;
