+++ backend/team6/src/test/java/programmers/team6/domain/auth/token/JwtConfigurationTests.java
@@ -1,4 +1,4 @@
-package programmers.team6.domain.auth.config;
+package programmers.team6.domain.auth.token;
 
 import static org.assertj.core.api.Assertions.*;
